requirements.txt
kafka-python
azure-storage-blob
pandas
matplotlib 